Support float frame_rate and related fixes
Convert frame_rate/frame_hz parameters from int to float across encoder, model, config, objective and output to allow non-integer frame rates. Add safe rounding when computing frame/sample sizes and audio context lengths (int(round(...))). Improve bin_times_to_frames to compute frames from float rates with rounding and minimum clamp. In EncoderMimi, avoid unnecessary resampling when frame rates match (math.isclose). Add _format_frame_rate to util and use a human-friendly frame_rate label in model file paths. Adjust GUI buffering to compute frame_samples from float frame_rate. Add a warning when no audio features are extracted. Minor example change: switch device to CPU and comment out output.update for testing. Also add required math import.

+++ src/maai/models/config.py
@@ -6,7 +6,7 @@ BIN_TIMES: list = [0.2, 0.4, 0.6, 0.8]
 @dataclass
 class VapConfig:
     sample_rate: int = 16000
-    frame_hz: int = 10
+    frame_hz: float = 10.0
     bin_times: List[float] = field(default_factory=lambda: BIN_TIMES)
 
     # Encoder (training flag)
